+++ tests/Telplin.Core.Tests/BindingTests.fs
@@ -70,7 +70,7 @@ let a (b:string list) : seq<int> = failwith "not implemented"
         """
 module A
 
-val a: b: string list -> seq<int>
+val a: b: string list -> int seq
 """
 
 [<Test>]
@@ -268,7 +268,7 @@ let (|Foo|) = function | 1 -> 2 | x -> x
         """
 module Hej
 
-val (|Foo|): (int -> int)
+val (|Foo|): int -> int
 """
 
 [<Test>]
@@ -494,7 +494,7 @@ let (|Two|) (DU.DU(one, two) as du) = two
 module Hej
 
 type DU = DU of one: string * two: int
-val (|Two|): DU -> int
+val (|Two|): du: DU -> int
 """
 
 [<Test>]
@@ -650,7 +650,7 @@ let f (a : string) (b : char) = fun (c : float) -> 2.0
         """
 module L
 
-val f: a: string -> b: char -> (float -> float)
+val f: a: string -> b: char -> c: float -> float
 """
 
 [<Test>]
@@ -664,7 +664,7 @@ let f (v : int -> int) (x : int -> int) : int -> int = fun z -> z + 1
         """
 module F
 
-val f: v: (int -> int) -> x: (int -> int) -> (int -> int)
+val f: v: (int -> int) -> x: (int -> int) -> z: int -> int
 """
 
 [<Test>]
@@ -678,7 +678,7 @@ let f (v : int -> int) (x : int -> int) = fun z -> z + 1
         """
 module F
 
-val f: v: (int -> int) -> x: (int -> int) -> (int -> int)
+val f: v: (int -> int) -> x: (int -> int) -> z: int -> int
 """
 
 [<Test>]
@@ -850,7 +850,7 @@ let intersperse separator (sequence: #seq<'a>) =
         """
 module FA
 
-val intersperse: separator: 'a -> sequence: #seq<'a> -> seq<'a>
+val intersperse: separator: 'a -> sequence: #('a seq) -> 'a seq
 """
 
 [<Test>]
@@ -1087,10 +1087,10 @@ module Telplin
 /// Optimized arrays equality. ~100x faster than `array1 = array2` on strings.
 /// ~2x faster for floats
 /// ~0.8x slower for ints
-val inline areEqual: xs: 'T[] -> ys: 'T[] -> bool when 'T: equality
+val inline areEqual: xs: 'T array -> ys: 'T array -> bool when 'T: equality
 /// check if subArray is found in the wholeArray starting
 /// at the provided index
-val inline isSubArray: subArray: 'T[] -> wholeArray: 'T[] -> index: int -> bool when 'T: equality
+val inline isSubArray: subArray: 'T array -> wholeArray: 'T array -> index: int -> bool when 'T: equality
 /// Returns true if one array has another as its subset from index 0.
 val startsWith: prefix: 'a array -> whole: 'a array -> bool when 'a: equality
 """
@@ -1131,7 +1131,7 @@ module Telplin
 type Node =
     abstract Children: int array
 
-val noa<'n when 'n :> Node> : n: 'n option -> Node array
+val noa: n: #Node option -> Node array
 """
 
 [<Test>]
@@ -1143,7 +1143,7 @@ module Telplin
 open System
 open System.Threading.Tasks
 
-let mapWithAdditionalDependenies
+let mapWithAdditionalDependencies
     (mapping: 'a -> 'b * #seq<#IDisposable>) 
     (value: Task<'a>) : Task<'b> =
     failwith "meh"
@@ -1154,8 +1154,8 @@ module Telplin
 open System
 open System.Threading.Tasks
 
-val mapWithAdditionalDependenies:
-    mapping: ('a -> 'b * #seq<#IDisposable>) -> value: Task<'a> -> Task<'b> when 'b1 :> IDisposable
+val mapWithAdditionalDependencies:
+    mapping: ('a -> 'b * #('b1 seq)) -> value: Task<'a> -> Task<'b> when 'b1 :> IDisposable
 """
 
 [<Test>]
@@ -1197,4 +1197,35 @@ module Telplin
 open System
 /// Reference equality.
 val inline (==): a: 'A -> b: 'B -> bool when 'A: not struct and 'B: not struct
+"""
+
+[<Test>]
+let ``property member with function return type`` () =
+    assertSignature
+        """
+module Telplin
+
+type Foo() =
+    member x.Bar = fun (i:int) -> ""
+"""
+        """
+module Telplin
+
+type Foo =
+    new: unit -> Foo
+    member Bar: (int -> string)
+"""
+
+[<Test>]
+let ``mutable value, 67`` () =
+    assertSignature
+        """
+module X
+
+let mutable lastDisplayContext : obj = null
+"""
+        """
+module X
+
+val mutable lastDisplayContext: obj
 """